chore(postprocessor): remove `cleanup.rs`, move logic to individual postprocessors
