+++ crates/common/src/ether/lexers/mod.rs
@@ -1,4 +1,3 @@
-pub mod cleanup;
 pub mod solidity;
 pub mod tokenize;
 pub mod yul;
